+++ src/FINDOLOGIC/Api/Exceptions/ConfigException.php
@@ -6,7 +6,7 @@ use RuntimeException;
 
 class ConfigException extends RuntimeException
 {
-    public function __construct($message = 'Invalid FindologicApi config.')
+    public function __construct($message = 'Invalid config supplied.')
     {
         parent::__construct($message);
     }
